changes for Bigint and target table master data

+++ src/specifications/service/genericFunction.ts
@@ -45,7 +45,7 @@ export class GenericFunction {
                 dbColumns[index] = 'VARCHAR'
             }
             else if (element?.type == "integer") {
-                dbColumns[index] = 'INT';
+                dbColumns[index] = 'BIGINT';
             }
             else if (element?.type == "json") {
                 dbColumns[index] = 'jsonb';
